\name{gwDist-package}
\alias{gwDist-package}
\alias{gwDist}
\docType{package}
\title{
\packageTitle{gwDist}
}
\description{
\packageDescription{gwDist}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{gwDist}
\packageIndices{gwDist}
}
\author{
\packageAuthor{gwDist}

Maintainer: \packageMaintainer{gwDist}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
